+++ core/src/main/scala/hammerlab/iterator/package.scala
@@ -1,0 +1,6 @@
+package hammerlab
+
+package object iterator
+  extends bulk
+     with scan
+     with sliding
