--- core/src/main/scala/hammerlab/iterators/package.scala
@@ -1,5 +1,0 @@
-package hammerlab
-
-package object iterators
-  extends bulk
-    with scan
