+++ core/src/main/scala/org/hammerlab/iterator/sliding/Sliding2Iterator.scala
@@ -1,8 +1,10 @@
 package org.hammerlab.iterator.sliding
 
+import hammerlab.iterator.macros.IteratorWrapper
 import org.hammerlab.iterator.SimpleBufferedIterator
 
-case class Sliding2Iterator[T](it: BufferedIterator[T]) {
+@IteratorWrapper
+class Sliding2Iterator[T](it: BufferedIterator[T]) {
   def sliding2Prev: Iterator[(Option[T], T)] =
     new Iterator[(Option[T], T)] {
       var prevOpt: Option[T] = None
@@ -29,10 +31,10 @@ case class Sliding2Iterator[T](it: BufferedIterator[T]) {
     }
 
   def sliding2(pad: T): Iterator[(T, T)] =
-    for {
-      (elem, succOpt) ← sliding2Opt
-    } yield
-      elem → succOpt.getOrElse(pad)
+    sliding2Opt.map {
+      case (elem, succOpt) ⇒
+        elem → succOpt.getOrElse(pad)
+    }
 
   def sliding2: Iterator[(T, T)] =
     new SimpleBufferedIterator[(T, T)] {
@@ -56,8 +58,8 @@ case class Sliding2Iterator[T](it: BufferedIterator[T]) {
     }
 }
 
-object Sliding2Iterator {
-  implicit def makeSliding2Iterator[T](it: Iterator[T]): Sliding2Iterator[T] = Sliding2Iterator(it.buffered)
-  implicit def makeSliding2Iterable[T](it: Iterable[T]): Sliding2Iterator[T] = Sliding2Iterator(it.iterator.buffered)
-  implicit def makeSliding2Array[T](it: Array[T]): Sliding2Iterator[T] = Sliding2Iterator(it.iterator.buffered)
-}
+//object Sliding2Iterator {
+//  implicit def makeSliding2Iterator[T](it: Iterator[T]): Sliding2Iterator[T] = Sliding2Iterator(it.buffered)
+//  implicit def makeSliding2Iterable[T](it: Iterable[T]): Sliding2Iterator[T] = Sliding2Iterator(it.iterator.buffered)
+//  implicit def makeSliding2Array[T](it: Array[T]): Sliding2Iterator[T] = Sliding2Iterator(it.iterator.buffered)
+//}
