+++ core/src/test/scala/org/hammerlab/iterator/bulk/BufferedCollectWhileTest.scala
@@ -1,6 +1,6 @@
 package org.hammerlab.iterator.bulk
 
-import hammerlab.iterators.bulk._
+import hammerlab.iterator.bulk._
 import org.hammerlab.iterator.EitherIteratorTest._
 import org.hammerlab.test.Suite
