+++ core/src/test/scala/org/hammerlab/iterator/sliding/Sliding2PadTest.scala
@@ -1,6 +1,6 @@
 package org.hammerlab.iterator.sliding
 
-import org.hammerlab.iterator.sliding.Sliding2Iterator._
+import hammerlab.iterator.sliding._
 import org.hammerlab.test.Suite
 
 class Sliding2PadTest
